--- a/src/samples/hello_world/program.c
+++ b/src/samples/hello_world/program.c
@@ -17,22 +17,22 @@ void get_root(http_request* request, hw_http_response* response, void* user_data
     hw_string body;
     hw_string keep_alive_name;
     hw_string keep_alive_value;
-    
+
     SETSTRING(status_code, HTTP_STATUS_200);
     hw_set_response_status_code(response, &status_code);
-    
+
     SETSTRING(content_type_name, "Content-Type");
-    
+
     SETSTRING(content_type_value, "text/html");
     hw_set_response_header(response, &content_type_name, &content_type_value);
-    
+
     SETSTRING(body, "hello world");
     hw_set_body(response, &body);
-    
+
     if (request->keep_alive)
     {
         SETSTRING(keep_alive_name, "Connection");
-        
+
         SETSTRING(keep_alive_value, "Keep-Alive");
         hw_set_response_header(response, &keep_alive_name, &keep_alive_value);
     }
@@ -40,13 +40,86 @@ void get_root(http_request* request, hw_http_response* response, void* user_data
     {
         hw_set_http_version(response, 1, 0);
     }
-    
+
+    hw_http_response_send(response, "user_data", response_complete);
+}
+
+void get_double_headers(http_request* request, hw_http_response* response, void* user_data)
+{
+    hw_string status_code;
+    hw_string content_type_name;
+    hw_string content_type_value;
+    hw_string body;
+    hw_string keep_alive_name;
+    hw_string keep_alive_value;
+
+    SETSTRING(status_code, HTTP_STATUS_200);
+    hw_set_response_status_code(response, &status_code);
+
+    SETSTRING(content_type_name, "Content-Type");
+
+    SETSTRING(content_type_value, "text/html");
+    hw_set_response_header(response, &content_type_name, &content_type_value);
+
+    SETSTRING(body, "hello world");
+    hw_set_body(response, &body);
+
+    if (request->keep_alive)
+    {
+        SETSTRING(keep_alive_name, "Connection");
+
+        SETSTRING(keep_alive_value, "Keep-Alive");
+        hw_add_response_header(response, &keep_alive_name, &keep_alive_value);
+        hw_add_response_header(response, &keep_alive_name, &keep_alive_value);
+    }
+    else
+    {
+        hw_set_http_version(response, 1, 0);
+    }
+
+    hw_http_response_send(response, "user_data", response_complete);
+}
+
+void get_overridden_headers(http_request* request, hw_http_response* response, void* user_data)
+{
+    hw_string status_code;
+    hw_string content_type_name;
+    hw_string content_type_value;
+    hw_string body;
+    hw_string keep_alive_name;
+    hw_string keep_alive_value;
+
+    SETSTRING(status_code, HTTP_STATUS_200);
+    hw_set_response_status_code(response, &status_code);
+
+    SETSTRING(content_type_name, "Content-Type");
+
+    SETSTRING(content_type_value, "text/html");
+    hw_set_response_header(response, &content_type_name, &content_type_value);
+
+    SETSTRING(body, "hello world");
+    hw_set_body(response, &body);
+
+    if (request->keep_alive)
+    {
+        SETSTRING(keep_alive_name, "Connection");
+
+        SETSTRING(keep_alive_value, "Keep-Alive");
+        hw_set_response_header(response, &keep_alive_name, &keep_alive_value);
+
+        SETSTRING(keep_alive_value, "close");
+        hw_set_response_header(response, &keep_alive_name, &keep_alive_value);
+    }
+    else
+    {
+        hw_set_http_version(response, 1, 0);
+    }
+
     hw_http_response_send(response, "user_data", response_complete);
 }
 
 int main(int args, char** argsv)
 {
-    char route[] = "/";
     configuration config;
     config.http_listen_address = "0.0.0.0";
     if (args > 1)
@@ -68,7 +141,10 @@ int main(int args, char** argsv)
     }
     /* hw_init_from_config("hello_world.conf"); */
     hw_init_with_config(&config);
-    hw_http_add_route(route, get_root, NULL);
+    hw_http_add_route("/", get_root, NULL);
+    hw_http_add_route("/double", get_double_headers, NULL);
+    hw_http_add_route("/override", get_overridden_headers, NULL);
+
     hw_http_open();
     return 0;
 }