The crucial bit here is that adding is additive only (i.e. you can double add a header) whereas setting will also add, but if it finds a header with the same name, it will simply update that existing header.

It will only override the first header it finds that matches, though, so if you double add a header, and then try to set the same header, it's only going to update the first.  I don't think that's a corner case worth dealing with, though, unless someone has a good use case to warrant scanning all headers outright.