@kellabyte 2015-07-21
Do you think we should keep `http_response.http_header` as an array or should we use `khash` like I did elsewhere? I may have done this because it was quicker at the time. I may have justified it because adding new headers should mostly be additive most of the time and an array is fast with small sets.