@tobz 2015-07-21
I guess, if we're trying to be objective, it boils down to: how much do we care about performance?  how much do we care about readability?

We can do it both ways.  Maybe we should just do some performance testing here.  We can achieve the same exact functionality either way.  If the performance is close enough, then maybe we think about readability?

(fwiw, what's your current benchmarking setup, at a high level? OS, client vs server setup, etc)