@kellabyte 2015-07-21
Make sure you don't benchmark on localhost. Use 2 separate machines for client and server.

This script is an example of a benchmark you can run on the client against Haywire:
https://github.com/kellabyte/Haywire/blob/master/benchmark/benchmark.sh

You can change the arguments to Wrk to change different workload scenarios like more connections or less pipelined requests.

You probably want to benchmark against small amounts of headers vs a decent amount of headers and different workloads of mostly additive vs 50% additive and 50% update or something.

Bottom line, we care about performance in every change we make. My goal with Haywire is to try and be the fastest HTTP server there is. Lofty goals but it keeps the focus on performance in mind :)