@kellabyte 2015-07-21
I believe this is what nginx is using to store headers: https://github.com/nginx/nginx/blob/220c36d6128bcf3e869fdd0f7767ead9a4222f08/src/core/ngx_list.c