@tobz 2015-07-23
Super preliminary: I can't actually measure the difference between sending along one "extra" header (excluding content type) vs sending along an extra four.  So that's good, I guess. :)  At least I now have the baseline to benchmark a khash approach against.

Unfortunately, I can't get Haywire to actually scale to more than one core.  Some sort of weird... under load it besides to lock up/freeze up and only one worker thread can actually do anything.

(Linux 3.14, Amazon Linux, GCC 4.8; tried with 2 threads, 4, 8, etc.)