@kellabyte 2015-07-23
Is this lockup happening because of the new changes or without them? Is there a crash happening that you might be able to trap with GDB?