@tobz 2015-07-27
Didn't get that far.  I was testing with big instances courtesy of my previous employer, so now I need to wait until I have more than just my desktop to run anymore benchmarks, wahhhh. :(

Obviously most anyone can pick this up, just won't be me for another week or two.