@nmdguerreiro 2015-12-17
I benchmarked this PR on two m4.xlarge machines on AWS. Haywire was setup to use 4 threads.

### Current master
```
[ec2-user@ip-172-31-5-244 wrk]$ ./wrk -c 100 -t 100 -d 5s -s pipelined_get.lua --latency http://172.31.5.65:8000
Running 5s test @ http://172.31.5.65:8000
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.31ms  135.91us  13.23ms   95.82%
    Req/Sec   763.07     90.24     2.46k    97.99%
  Latency Distribution
     50%    1.32ms
     75%    1.35ms
     90%    1.38ms
     99%    1.46ms
  385766 requests in 5.10s, 58.13MB read
Requests/sec:  75647.78
Transfer/sec:     11.40MB
```

### @tobz's changes - `GET /``
```
[ec2-user@ip-172-31-5-244 wrk]$ ./wrk -c 100 -t 100 -d 5s -s pipelined_get.lua --latency http://172.31.5.65:8000
Running 5s test @ http://172.31.5.65:8000
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.19ms  154.72us  16.21ms   92.83%
    Req/Sec   835.31    104.96     2.27k    97.17%
  Latency Distribution
     50%    1.21ms
     75%    1.27ms
     90%    1.32ms
     99%    1.39ms
  385649 requests in 5.10s, 58.11MB read
  Socket errors: connect 0, read 5, write 0, timeout 0
  Non-2xx or 3xx responses: 175
Requests/sec:  75617.78
Transfer/sec:     11.39MB
```

### @tobz's changes - `GET /double`
```
[ec2-user@ip-172-31-5-244 wrk]$ ./wrk -c 100 -t 100 -d 5s -s pipelined_get.lua --latency http://172.31.5.65:8000
Running 5s test @ http://172.31.5.65:8000
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.16ms  205.87us  22.33ms   87.71%
    Req/Sec     0.86k   118.57     1.83k    84.24%
  Latency Distribution
     50%    1.18ms
     75%    1.27ms
     90%    1.32ms
     99%    1.40ms
  384622 requests in 5.10s, 74.26MB read
  Socket errors: connect 0, read 57, write 0, timeout 0
  Non-2xx or 3xx responses: 2060
Requests/sec:  75435.91
Transfer/sec:     14.56MB
```

### @tobz's changes - `GET /override`
```
[ec2-user@ip-172-31-5-244 wrk]$ ./wrk -c 100 -t 100 -d 5s -s pipelined_get.lua --latency http://172.31.5.65:8000
Running 5s test @ http://172.31.5.65:8000
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.12ms  161.15us  10.28ms   75.53%
    Req/Sec     0.88k   126.64     2.26k    78.61%
  Latency Distribution
     50%    1.11ms
     75%    1.24ms
     90%    1.32ms
     99%    1.40ms
  385123 requests in 5.10s, 74.10MB read
  Socket errors: connect 0, read 28, write 0, timeout 0
  Non-2xx or 3xx responses: 909
Requests/sec:  75514.61
Transfer/sec:     14.53MB
```

### @tobz's changes - `GET /mixandmatch`
```
[ec2-user@ip-172-31-5-244 wrk]$ ./wrk -c 100 -t 100 -d 5s -s pipelined_get.lua --latency http://172.31.5.65:8000
Running 5s test @ http://172.31.5.65:8000
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.26ms  134.63us  10.96ms   94.61%
    Req/Sec   793.59     82.68     1.96k    97.84%
  Latency Distribution
     50%    1.27ms
     75%    1.31ms
     90%    1.35ms
     99%    1.42ms
  384869 requests in 5.10s, 117.64MB read
  Socket errors: connect 0, read 28, write 0, timeout 0
  Non-2xx or 3xx responses: 1105
Requests/sec:  75470.10
Transfer/sec:     23.07MB
```

As you can see, results are pretty much identical, just that having more headers is a bit slower as is expected.

You'll see a number of non 2xx/3xx responses, but that is not caused by this commit, but by a race condition that exists on `route_compare_method.c`. I opened #78 to fix it.