@tobz 2015-12-17
@nmdguerreiro thanks for doing some consistent tests on this!