@kellabyte 2015-12-19
I did some more aggressive benchmarking.

# Master branch
```
../bin/wrk/wrk --script ./pipelined_get.lua --latency -d 10s -t 40 -c 512 http://server:8000 -- 32
Running 10s test @ http://server:8000
  40 threads and 512 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    16.10ms   12.74ms 101.92ms   55.55%
    Req/Sec    64.43k    20.61k  177.88k    70.60%
  Latency Distribution
     50%   14.12ms
     75%   27.06ms
     90%   34.87ms
     99%   41.82ms
  25786650 requests in 10.10s, 3.79GB read
Requests/sec: 2553250.20
Transfer/sec:    384.73MB
```

# This PR branch
```
./bin/wrk/wrk --script ./pipelined_get.lua --latency -d 10s -t 40 -c 512 http://server:8000 -- 32
Running 10s test @ http://server:8000
  40 threads and 512 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    18.74ms   12.09ms 119.86ms   57.91%
    Req/Sec    23.28k    11.42k   70.10k    67.40%
  Latency Distribution
     50%   18.58ms
     75%   29.01ms
     90%   35.31ms
     99%   39.87ms
  9006443 requests in 10.10s, 1.32GB read
  Socket errors: connect 0, read 29925, write 0, timeout 0
  Non-2xx or 3xx responses: 1642156
Requests/sec: 891678.30
Transfer/sec:    133.47MB
```

Should we test this with fix #78?